Create test_place_stone() and test_remove_stone()
I also added the cool TEST ZONE text that divides the code so I can just look at the code itself with no distractions.

Everything else is just cleanup or removal of unnecessary complexity at this point of development.

A learning point to take away from this is start by tackling what is immediately in front of you, not what is down the road.